requirements.txt
yfinance
pandas
lxml
html5lib
line-bot-sdk